KinematicsConfig

+++ module/actuation/KinematicsConfiguration/src/KinematicsConfiguration.cpp
@@ -52,23 +52,23 @@ namespace module::actuation {
     }
 
     void KinematicsConfiguration::configureLeg(KinematicsModel& model, const YAML::Node& objLeg) {
-        Eigen::Vector3f leg_hipOffset = objLeg["hip_offset"].as<Expression>();
+        Eigen::Vector3d leg_hipOffset = objLeg["hip_offset"].as<Expression>();
         model.leg.HIP_OFFSET_X        = leg_hipOffset.x();
         model.leg.HIP_OFFSET_Y        = leg_hipOffset.y();
         model.leg.HIP_OFFSET_Z        = leg_hipOffset.z();
 
-        model.leg.UPPER_LEG_LENGTH = objLeg["upper_leg_length"].as<float>();
-        model.leg.LOWER_LEG_LENGTH = objLeg["lower_leg_length"].as<float>();
+        model.leg.UPPER_LEG_LENGTH = objLeg["upper_leg_length"].as<double>();
+        model.leg.LOWER_LEG_LENGTH = objLeg["lower_leg_length"].as<double>();
 
-        model.leg.HEEL_LENGTH = objLeg["heel_length"].as<float>();
+        model.leg.HEEL_LENGTH = objLeg["heel_length"].as<double>();
 
-        model.leg.FOOT_CENTRE_TO_ANKLE_CENTRE = objLeg["foot_centre_to_ankle_centre"].as<float>();
+        model.leg.FOOT_CENTRE_TO_ANKLE_CENTRE = objLeg["foot_centre_to_ankle_centre"].as<double>();
 
         const auto& objFoot   = objLeg["foot"];
-        model.leg.FOOT_WIDTH  = objFoot["width"].as<float>();
-        model.leg.FOOT_HEIGHT = objFoot["height"].as<float>();
-        model.leg.FOOT_LENGTH = objFoot["length"].as<float>();
-        model.leg.TOE_LENGTH  = objFoot["toe_length"].as<float>();
+        model.leg.FOOT_WIDTH  = objFoot["width"].as<double>();
+        model.leg.FOOT_HEIGHT = objFoot["height"].as<double>();
+        model.leg.FOOT_LENGTH = objFoot["length"].as<double>();
+        model.leg.TOE_LENGTH  = objFoot["toe_length"].as<double>();
 
         model.leg.LENGTH_BETWEEN_LEGS = 2.0 * model.leg.HIP_OFFSET_Y;
 
@@ -84,25 +84,25 @@ namespace module::actuation {
     void KinematicsConfiguration::configureHead(KinematicsModel& model, const YAML::Node& objHead) {
         model.head.CAMERA_DECLINATION_ANGLE_OFFSET = objHead["camera_declination_angle_offset"].as<Expression>();
 
-        Eigen::Vector3f head_neckToCamera  = objHead["neck_to_camera"].as<Expression>();
+        Eigen::Vector3d head_neckToCamera  = objHead["neck_to_camera"].as<Expression>();
         model.head.NECK_TO_CAMERA_X        = head_neckToCamera.x();
         model.head.NECK_TO_CAMERA_Y        = head_neckToCamera.y();
         model.head.NECK_TO_CAMERA_Z        = head_neckToCamera.z();
-        model.head.INTERPUPILLARY_DISTANCE = objHead["ipd"].as<float>();
+        model.head.INTERPUPILLARY_DISTANCE = objHead["ipd"].as<double>();
 
         const auto& objNeck = objHead["neck"];
 
-        model.head.NECK_LENGTH = objNeck["length"].as<float>();
+        model.head.NECK_LENGTH = objNeck["length"].as<double>();
 
-        Eigen::Vector3f neck_basePositionFromOrigin = objNeck["base_position_from_origin"].as<Expression>();
+        Eigen::Vector3d neck_basePositionFromOrigin = objNeck["base_position_from_origin"].as<Expression>();
         model.head.NECK_BASE_POS_FROM_ORIGIN_X      = neck_basePositionFromOrigin.x();
         model.head.NECK_BASE_POS_FROM_ORIGIN_Y      = neck_basePositionFromOrigin.y();
         model.head.NECK_BASE_POS_FROM_ORIGIN_Z      = neck_basePositionFromOrigin.z();
 
         const auto& objHeadMovementLimits = objHead["limits"];
 
-        Eigen::Vector2f headMovementLimits_yaw   = objHeadMovementLimits["yaw"].as<Expression>();
-        Eigen::Vector2f headMovementLimits_pitch = objHeadMovementLimits["pitch"].as<Expression>();
+        Eigen::Vector2d headMovementLimits_yaw   = objHeadMovementLimits["yaw"].as<Expression>();
+        Eigen::Vector2d headMovementLimits_pitch = objHeadMovementLimits["pitch"].as<Expression>();
         model.head.MIN_YAW                       = headMovementLimits_yaw.x();
         model.head.MAX_YAW                       = headMovementLimits_yaw.y();
         model.head.MIN_PITCH                     = headMovementLimits_pitch.x();
@@ -114,21 +114,21 @@ namespace module::actuation {
         const auto& objUpperArm = objArm["upper_arm"];
         const auto& objLowerArm = objArm["lower_arm"];
 
-        model.arm.DISTANCE_BETWEEN_SHOULDERS = objArm["distance_between_shoulders"].as<float>();
-        Eigen::Vector2f shoulderOffset       = objShoulder["offset"].as<Expression>();
+        model.arm.DISTANCE_BETWEEN_SHOULDERS = objArm["distance_between_shoulders"].as<double>();
+        Eigen::Vector2d shoulderOffset       = objShoulder["offset"].as<Expression>();
         model.arm.SHOULDER_X_OFFSET          = shoulderOffset.x();
         model.arm.SHOULDER_Z_OFFSET          = shoulderOffset.y();
-        model.arm.SHOULDER_LENGTH            = objShoulder["length"].as<float>();
-        model.arm.SHOULDER_WIDTH             = objShoulder["width"].as<float>();
-        model.arm.SHOULDER_HEIGHT            = objShoulder["height"].as<float>();
+        model.arm.SHOULDER_LENGTH            = objShoulder["length"].as<double>();
+        model.arm.SHOULDER_WIDTH             = objShoulder["width"].as<double>();
+        model.arm.SHOULDER_HEIGHT            = objShoulder["height"].as<double>();
 
-        model.arm.UPPER_ARM_LENGTH     = objUpperArm["length"].as<float>();
-        Eigen::Vector2f upperArmOffset = objUpperArm["offset"].as<Expression>();
+        model.arm.UPPER_ARM_LENGTH     = objUpperArm["length"].as<double>();
+        Eigen::Vector2d upperArmOffset = objUpperArm["offset"].as<Expression>();
         model.arm.UPPER_ARM_Y_OFFSET   = upperArmOffset.x();
         model.arm.UPPER_ARM_X_OFFSET   = upperArmOffset.y();
 
-        model.arm.LOWER_ARM_LENGTH     = objLowerArm["length"].as<float>();
-        Eigen::Vector2f lowerArmOffset = objLowerArm["offset"].as<Expression>();
+        model.arm.LOWER_ARM_LENGTH     = objLowerArm["length"].as<double>();
+        Eigen::Vector2d lowerArmOffset = objLowerArm["offset"].as<Expression>();
         model.arm.LOWER_ARM_Y_OFFSET   = lowerArmOffset.x();
         model.arm.LOWER_ARM_Z_OFFSET   = lowerArmOffset.y();
     }
